refactor: ImageService -> ImageUploader 클래스 명 변경 및 utils 패키지로 이동
