+++ back-end/legeno-around-here/src/main/java/wooteco/team/ittabi/legenoaroundhere/utils/ImageUploader.java
@@ -1,4 +1,4 @@
-package wooteco.team.ittabi.legenoaroundhere.service;
+package wooteco.team.ittabi.legenoaroundhere.utils;
 
 import java.io.IOException;
 import java.security.MessageDigest;
@@ -8,7 +8,7 @@ import javax.xml.bind.DatatypeConverter;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.tika.Tika;
-import org.springframework.stereotype.Service;
+import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 import wooteco.team.ittabi.legenoaroundhere.aws.S3Uploader;
@@ -21,9 +21,9 @@ import wooteco.team.ittabi.legenoaroundhere.exception.NotImageMimeTypeException;
 
 @Slf4j
 @Transactional
-@Service
+@Component
 @AllArgsConstructor
-public class ImageService {
+public class ImageUploader {
 
     public static final String IMAGE_TYPE = "image";
     public static final String IMAGE_DIR = "posts/images/";
@@ -32,7 +32,7 @@ public class ImageService {
     private final S3Uploader s3Uploader;
     private final IAuthenticationFacade authenticationFacade;
 
-    public Image upload(MultipartFile multipartFile) {
+    public Image uploadImage(MultipartFile multipartFile) {
         User user = (User) authenticationFacade.getPrincipal();
         validateImage(multipartFile);
         String imageUrl = s3Uploader
